feat: CRUD operations now checks authentication

+++ controllers/contacts/removeContact.js
@@ -1,8 +1,10 @@
+const checkIfOwner = require("../../helpers/checkIfOwner");
 const createErrorMessage = require("../../helpers/createErrorMessage");
 const Contact = require("../../models/contact");
 
 const removeContact = async (req, res, next) => {
   try {
+    await checkIfOwner(req);
     const response = await Contact.findByIdAndDelete(req.params.contactId);
     if (!response) {
       throw createErrorMessage(404);
